Review Changes
This commit updates the following:
1. Uses a CRTP based singleton solution as opposed to a generic
single-singleton per type solution. This allows multiple singleton
instances of the same type.
2. Add register.hpp which includes all the registration functions.
3. Put system_detector_ in core.cpp.
4. Update naming scheme of enum class.
