+++ include/charmlite/core/destination.hpp
@@ -37,12 +37,12 @@ namespace cmk {
         friend struct message;
 
         destination(void)
-          : kind_(destination_kind::kInvalid)
+          : kind_(destination_kind::Invalid)
         {
         }
 
         destination(callback_id_t id, int pe)
-          : kind_(destination_kind::kCallback)
+          : kind_(destination_kind::Callback)
         {
             new (&(this->impl_.callback_fn_))
                 s_callback_fn_{.id = id, .pe = pe};
@@ -50,7 +50,7 @@ namespace cmk {
 
         destination(const collection_index_t& collection,
             const chare_index_t& chare, entry_id_t entry)
-          : kind_(destination_kind::kEndpoint)
+          : kind_(destination_kind::Endpoint)
         {
             new (&(this->impl_.endpoint_)) s_endpoint_{.collection = collection,
                 .chare = chare,
@@ -63,13 +63,13 @@ namespace cmk {
 
         inline s_callback_fn_& callback_fn(void)
         {
-            CmiAssert(this->kind_ == destination_kind::kCallback);
+            CmiAssert(this->kind_ == destination_kind::Callback);
             return this->impl_.callback_fn_;
         }
 
         inline s_endpoint_& endpoint(void)
         {
-            CmiAssert(this->kind_ == destination_kind::kEndpoint);
+            CmiAssert(this->kind_ == destination_kind::Endpoint);
             return this->impl_.endpoint_;
         }
 
@@ -82,9 +82,9 @@ namespace cmk {
         {
             switch (this->kind_)
             {
-            case destination_kind::kCallback:
+            case destination_kind::Callback:
                 return (this->impl_.callback_fn_.pe == cmk::all::pes);
-            case destination_kind::kEndpoint:
+            case destination_kind::Endpoint:
             {
                 auto& ep = this->impl_.endpoint_;
                 return ep.bcast ||
@@ -97,7 +97,7 @@ namespace cmk {
 
         operator bool(void) const
         {
-            return !(this->kind_ == destination_kind::kInvalid);
+            return !(this->kind_ == destination_kind::Invalid);
         }
 
         operator std::string(void) const
@@ -106,14 +106,14 @@ namespace cmk {
             ss << "destination(";
             switch (kind_)
             {
-            case destination_kind::kCallback:
+            case destination_kind::Callback:
             {
                 auto& cb = this->impl_.callback_fn_;
                 ss << "cb=" << cb.id << ",";
                 ss << "pe=" << cb.pe;
                 break;
             }
-            case destination_kind::kEndpoint:
+            case destination_kind::Endpoint:
             {
                 auto& ep = this->impl_.endpoint_;
                 ss << (std::string) ep.collection << ",";
